requirements.txt
bcrypt
Flask
gunicorn
psycopg2-binary
python-dotenv
requests
cloudinary